Expose FFT demod context and reuse
