+++ tests/CMakeLists.txt
@@ -155,12 +155,12 @@ lora_add_test(test_full_chain_compare
   WORKING_DIR ${CMAKE_SOURCE_DIR})
 
 lora_add_test(test_fixed_float_float
-  SOURCES test_fixed_float_equiv.c ../src/lora_mod.c ../src/lora_fft_demod.c ../src/lora_fft_demod_ctx.c ../src/lora_utils.c ../src/lora_fft.c ../src/q15_to_cf.c
+  SOURCES test_fixed_float_equiv.c ../src/lora_mod.c ../src/lora_fft_demod.c ../src/lora_utils.c ../src/lora_fft.c ../src/q15_to_cf.c
   LIBS m
   DEFINITIONS LORA_LITE_FIXED_POINT)
 
 lora_add_test(test_fixed_float_fixed
-  SOURCES test_fixed_float_equiv.c ../src/lora_mod.c ../src/lora_fft_demod.c ../src/lora_fft_demod_ctx.c ../src/lora_utils.c ../src/lora_fft.c ../src/q15_to_cf.c
+  SOURCES test_fixed_float_equiv.c ../src/lora_mod.c ../src/lora_fft_demod.c ../src/lora_utils.c ../src/lora_fft.c ../src/q15_to_cf.c
   LIBS m
   DEFINITIONS LORA_LITE_FIXED_POINT)
